Reorganize, docs
