+++ example/team/metadata.md
@@ -14,5 +14,40 @@ table.
 
 {{ adr::header_table() }}
 
-- [ ] task 1
-- [x] task 2
+## Referencing metadata
+
+You can reference metadata in the following way:
+
+```markdown
+---
+metadata:
+  author: James
+---
+
+# Title
+
+{{ '{{ metadata(path="author") }}'}}
+```
+
+As of `1.0.4`, metadata can also be referenced via the `fm` variable:
+
+```markdown
+---
+metadata:
+  author: James
+---
+
+# Title
+
+{{ '{{ fm.metadata.author }}'}}
+```
+
+This allows some extra work to be done in the templating engine. For instance,
+you can print out all of the metadata like this:
+
+
+| key | value |
+| --- | --- |
+{% for key, value in fm.metadata -%}
+| {{ key}} | {{value}} |
+{% endfor %}
